ultimo añadidos

+++ database/migrations/2025_02_18_172521_add_category_id_to_species_table.php
@@ -10,19 +10,25 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('species', function (Blueprint $table) {
-        $table->unsignedBigInteger('category_id')->nullable()->after('location');
-        $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
-    });
-}
+    {
+        Schema::table('species', function (Blueprint $table) {
+            if (!Schema::hasColumn('species', 'category_id')) {
+                $table->unsignedBigInteger('category_id')->nullable()->after('location');
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('species', function (Blueprint $table) {
-        $table->dropForeign(['category_id']);
-        $table->dropColumn('category_id');
-        
-    });
-}
+    /**
+     * Reverse the migrations.
+     */
+    public function down()
+    {
+        Schema::table('species', function (Blueprint $table) {
+            if (Schema::hasColumn('species', 'category_id')) {
+                $table->dropForeign(['category_id']);
+                $table->dropColumn('category_id');
+            }
+        });
+    }
 };
